cook policy-keystore into pe image
OOM team reported issues with ONAP installation in
terms of config map sizes.   After discussion with
OOM team, one solution is to cook it in the docker
image itself.   This will solve the size issue on
the OOM side.

Issue-ID: POLICY-747

+++ packages/docker/src/main/docker/do-start.sh
@@ -64,9 +64,10 @@ else
 
 	. /opt/app/policy/etc/profile.d/env.sh
 
-	# install policy keystore
-	mkdir -p $POLICY_HOME/etc/ssl
-	cp config/policy-keystore $POLICY_HOME/etc/ssl
+	if [[ -f config/policy-keystore ]]; then
+	    # install policy keystore
+	    cp config/policy-keystore $POLICY_HOME/etc/ssl
+	fi
 
 	if [[ -f config/$container-tweaks.sh ]] ; then
 		# file may not be executable; running it as an
